requirements.txt
google-generativeai>=0.3.0
wikipedia-api>=0.6.0
python-dotenv>=1.0.0
streamlit>=1.28.0
